Updated module augmentation for clipboard package.
