+++ packages/ckeditor5-clipboard/src/clipboardpipeline.ts
@@ -441,9 +441,3 @@ export interface ViewDocumentClipboardOutputEventData {
 	 */
 	method: 'copy' | 'cut' | 'dragstart';
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ ClipboardPipeline.pluginName ]: ClipboardPipeline;
-	}
-}
