+++ packages/ckeditor5-clipboard/src/index.ts
@@ -27,3 +27,5 @@ export type {
 	ViewDocumentCopyEvent,
 	ViewDocumentCutEvent
 } from './clipboardobserver';
+
+import './augmentation';
