+++ packages/ckeditor5-clipboard/src/pasteplaintext.ts
@@ -101,9 +101,3 @@ function isPlainTextFragment( documentFragment: DocumentFragment, schema: Schema
 
 	return Array.from( child.getAttributeKeys() ).length == 0;
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ PastePlainText.pluginName ]: PastePlainText;
-	}
-}
